fix: keep image keying kernel sharp-free (#4654)
